20170301
last update 01/03/2017 (2017030101)
- new tables were created to render
the recylcle bin functional
- new settings were created to permit users
to personnalise their mails and reminders
- if the setting
'via_activitydeletion' is true we still pemit users to delete recordings
but only in moodle,
they continue to existe in Via.
- a new status
exists for duplicated or restored activities; these are now unplanned
activities and do not yet existe in Via.
- PHP 7 was installed and used
to validate the code.
- many bug fixes included.